backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.16.0
environs==14.1.1
gunicorn==23.0.0
idna==3.10
marshmallow==3.26.1
packaging==24.2
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
